+++ src/gh-slugify.js
@@ -61,7 +61,9 @@ export default class GithubSlugify {
       uniqueSlug = `${slug}-${this.slugs[slug]}`;
     }
     this.slugs[slug] += 1;
-    return uniqueSlug;
+
+    // Remove any trailing hyphens before returning
+    return uniqueSlug.replace(/-$/g, '');
   }
 
   /**