+++ src/index.js
@@ -2,6 +2,6 @@ import ReactMarkdownGithub from './component.js';
 import GithubSlugify from './gh-slugify.js';
 
 export {
-  ReactMarkdownGithub,
+  ReactMarkdownGithub as default,
   GithubSlugify
 };